requirements.txt
langchain==0.0.267
langchain-community
psycopg[binary]
fastapi
uvicorn
jinja2
python-dotenv
pydantic
sentence-transformers
tqdm
numpy
requests
tenacity